[DNM into 9.0]Utils: Introduce method to determine countries that use Fahrenheit
This one is a weird one because up until a few days ago, I didn't think
I would need. Basically what this does is check the device's MCC and bounce
it off an array of MCC's that belong to countries that use fahrenheit.

- If no MCC is found (whether is a WiFi only tablet or the user has their phone
in airplane mode), the method returns false (Celsius). Celsius is used by 99 %
of all people so is not a bad option.

Inspired by Henrique Silva <jhenrique09.mcz@hotmail.com> and his commit below
https://gerrit.pixelexperience.org/#/c/frameworks_base/+/1606/

We're going to end up using this for two things, our battery temp thingy in
Settings and also for what Henrique used it for, the keyguard weather default.

 * martinusbe: got hyperactive while merging this, no idea why, i should have done
  all this in GzospUtils. Blame old age, CBD, cancer, ...anything but me
  But do not merge in stable branch, thx

Change-Id: I49c7a397c794e7852ab0be4bd3182710fa3b19b0