Set scrollview on the power menu
No, you need this for if and when we decide to add more items to the power menu and the
density is too high. Previously if you had more than 5 items, it would cut you off. So
you either had to decide which 5 items you wanted or deal with the jank. That's no longer
the case.

Updated for 10

Change-Id: I906462458858318709b3ccd4d0b95fa724098505